Added Item.cs

+++ GridRL/Creature.cs
@@ -18,7 +18,7 @@ namespace GridRL {
         /// <param name="image"> The image used for this creature sprite</param>
         /// <param name="x"> The creature's X position on the world data. </param>
         /// <param name="y"> The creature's Y position on the world data.</param>
-        public Creature(Image image, float x, float y) : base(image, x, y) { }
+        public Creature(Image image, int x, int y) : base(image, x, y) { }
 
 
         /* Properties */
